aeroflot_backend/requirements.txt
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.19
django-cors-headers==4.7.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
outcome==1.3.0.post0
psycopg2-binary==2.9.10
PySocks==1.7.1
requests==2.32.3
selenium==4.29.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
